Bug fix of project core
