+++ package.json
@@ -4,7 +4,9 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build": "webpack --mode production",
+    "serve": "webpack-dev-server --mode production",
+    "serve:dev": "webpack-dev-server --mode development"
   },
   "keywords": [],
   "author": "",
@@ -18,7 +20,7 @@
     "@babel/core": "^7.12.10",
     "@babel/preset-env": "^7.12.10",
     "@webpack-cli/serve": "^1.1.0",
-    "autoprefixer": "^10.1.0",
+    "autoprefixer": "^9.8.6",
     "babel-loader": "^8.2.2",
     "clean-webpack-plugin": "^3.0.0",
     "css-loader": "^4.3.0",
