+++ public/index.html
@@ -2,8 +2,8 @@
 <html>
 	<head>
 		<meta charset="UTF-8">
-		<meata name="viewport" content="width=device-width, initial-scale=1.0">
-		<tilte>renome_vue</tilte>
+		<meta name="viewport" content="width=device-width, initial-scale=1.0">
+		<title>renome_vue</title>
 	</head>
 	<body>
 		<div id="app"></div>
