+++ src/App.vue
@@ -33,6 +33,6 @@ export default {
 }
 </script>
 
-<style scoped>
+<style>
 
 </style>
