+++ src/components/molecules/galleryRow.vue
@@ -1,0 +1,11 @@
+<template>
+
+</template>
+
+<script>
+
+</script>
+
+<style>
+
+</style>
