+++ src/components/molecules/navigationExpandableItem.vue
@@ -1,0 +1,11 @@
+<template>
+  
+</template>
+
+<script>
+
+</script>
+
+<style>
+
+</style>
