+++ src/components/molecules/navigationSubmenu.vue
@@ -1,0 +1,11 @@
+<template>
+  
+</template>
+
+<script>
+
+</script>
+
+<style>
+
+</style>
