+++ src/components/organisms/building.vue
@@ -1,0 +1,22 @@
+<template>
+	<p></p>
+</template>
+
+<script>
+	export default {
+		name: 'building',
+		components: {
+			
+		}
+	}
+</script>
+
+<style scoped>
+	.building {
+		width: 100%;
+		height: auto;
+		position: relative;
+		display: flex;
+		flex-direction: column;
+	}
+</style>
