+++ src/components/organisms/gallery.vue
@@ -1,0 +1,22 @@
+<template>
+	<p></p>
+</template>
+
+<script>
+	export default {
+		name: 'gallery',
+		components: {
+			
+		}
+	}
+</script>
+
+<style scoped>
+	.gallery {
+		display: flex;
+		flex-direction: column;
+		width: 100vw;
+		background-color: var(--lightGray);
+		padding: 70px 0;
+	}
+</style>
