+++ src/components/organisms/home.vue
@@ -1,0 +1,22 @@
+<template>
+	<p></p>
+</template>
+
+<script>
+	export default {
+		name: 'home',
+		components: {
+			
+		}
+	}
+</script>
+
+<style>
+	.home {
+		position: relative;
+		margin: 60px 0 70px;
+		height: 140vw;
+		width: 100vw;
+		overflow: hidden;
+	}
+</style>
