--- src/organisms/about.vue
@@ -1,22 +1,0 @@
-<template>
-	<p></p>
-</template>
-
-<script>
-export default {
-  name: 'about',
-  components: {
-    
-  }
-}
-</script>
-
-<style scoped>
-.about {
-	display: flex;
-	box-sizing: border-box;
-	flex-direction: column;
-	width: 100%;
-	padding: 30px 10px 70px;
-}
-</style>
