--- src/organisms/gallery.vue
@@ -1,22 +1,0 @@
-<template>
-	<p></p>
-</template>
-
-<script>
-export default {
-	name: 'gallery',
-	components: {
-    
-  }
-}
-</script>
-
-<style scoped>
-.gallery {
-	display: flex;
-	flex-direction: column;
-	width: 100vw;
-	background-color: var(--lightGray);
-	padding: 70px 0;
-}
-</style>
