--- src/organisms/tail.vue
@@ -1,24 +1,0 @@
-<template>
-	<p></p>
-</template>
-
-<script>
-export default {
-	name: 'tail',
-	components: {
-    
-  }
-}
-</script>
-
-<style scoped>
-.footer {
-	position: relative;
-	display: flex;
-	box-sizing: border-box;
-	flex-direction: column;
-	background-color: var(--gray);
-	padding: 20px 10px;
-	width: 100vw;
-}
-</style>
